Move semistructured query unit tests to the right class
